Add parser modules for ARG Database graphs for isomorphism benchmarks
This commit adds a parser for the unlabeled graph files in the ARG
database: https://mivia.unisa.it/datasets/graph-database/arg-database/
which we'll be using for benchmarking the graph isomorphism functions. I
only implemented the unlabelled graphs since it was simpler. If we need
labels this can be expanded. The file format is documented here:
https://mivia.unisa.it/datasets/graph-database/arg-database/documentation/

It's worth noting that the graph database is comprised of nested zip
files for different sets of graphs and the function added for each
library variant here is only for reading a single file. Either we'll
have to update the run script to unzip the graphs we want or use
python's zipfile module (https://docs.python.org/3/library/zipfile.html)
to automate extraction of graphs on top of this function in the
benchmark setup.