Re-run benchmarks and regenerate data
I had to install a newer python version and some other system library
updates on the benchmark machine for some unrelated tasks. Since we're
not finished collecting data for other benchmarks I went ahead and
regenerated the entire data set after the update. This commit contains
the new data and updates the python version in the README.