Reduce workload for isomorphism benchmarks
The isomorphism problems being benchmarked were generally too slow to be
useful. For most of the graphs it was taking hours to run a single graph
for most of the database and it wasn't useful for benchmarking. This
commit restricts the data set to just those that execute quickly
(typically taking < 1 sec). This lets us rerun the benchmarks quickly.