Adjust igraph graph creation benchmark
It was pointed out that the graph creation benchmark for igraph wasn't
written optimally for performance. It was using networkx to create a
graph and then converting that to an igraph graph. This was done
originally because this was significantly faster than calling
add_edge() in a loop (which for igraph was taking hours) like we do for
all the other graph libraries. Apparently igraph is very inefficient on
graph modification but more performant for creation if it's static all
at once. This commit adjust the benchmarks to create an intermediate
edge list first and then add all the edges at once. This is much faster
than using networkx as was done previously.